+++ Journet-Web/Home/main.html
@@ -54,7 +54,6 @@
       <h3>여행 기록 한 눈에 보기</h3>
       <div class="map_view">
         <script src="https://unpkg.com/@supabase/supabase-js@2"></script>
-        <script src="/Journet/Journet-Web/DB/Supabase/Mappins.js"></script>
         <div id="map"></div>
       </div>
     </div>
@@ -351,7 +350,7 @@
 
   <!-- Kakao map API here -->
   <script src="https://dapi.kakao.com/v2/maps/sdk.js?appkey=564115dcd7ead3ed0d36372d5583c5b1"></script>
-  <script src="../API/KakaoMap.js"></script>
+  <script src="../DB/Supabase/Supabase.js"></script>
 
   <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
